+++ stories/index.tsx
@@ -21,6 +21,7 @@ import '../src/layout/section/mediaContentSection/story'
 import '../src/layout/section/columnedContentSection/story'
 import '../src/layout/section/illustratedSection/story'
 import '../src/layout/section/itemsSection/story'
+import '../src/layout/section/mediaSection/story'
 
 // Pages
 import '../src/pages/messaging/inbox.story'